refactor: rename session-named TypeScript files to conversation
