+++ assistant/src/__tests__/handlers-user-message-approval-consumption.test.ts
@@ -143,7 +143,7 @@ mock.module("../util/logger.js", () => ({
   }),
 }));
 
-import { handleConfirmationResponse } from "../daemon/handlers/sessions.js";
+import { handleConfirmationResponse } from "../daemon/handlers/conversations.js";
 
 interface TestSession {
   messages: Array<{ role: string; content: unknown[] }>;
