+++ assistant/src/daemon/handlers/conversation-history.ts
