+++ assistant/src/daemon/handlers/conversations.ts
