+++ assistant/src/daemon/message-types/conversations.ts
